+++ platformio/__main__.py
@@ -41,10 +41,10 @@ class PlatformioCLI(click.MultiCommand):  # pylint: disable=R0904
     def _handle_obsolate_command(name):
         if name in ("install", "list", "search", "show", "uninstall"):
             click.secho(
-                "Warning! `platformio %s` command is obsoleted and will be "
+                "Warning! `platformio %s` command is deprecated and will be "
                 "removed in the next release! Please use "
                 "`platformio platforms %s` instead." % (name, name),
-                fg="red"
+                fg="yellow"
             )
             from platformio.commands import platforms
             return getattr(platforms, "platforms_" + name)